Info
As a programmer, I have recently completed a project that involves creating a functional calculator using React, HTML, CSS, and JavaScript. Through this project, I have gained a deep understanding of the fundamental concepts of these programming languages, including the use of built-in libraries, effective page styling techniques, and the implementation of various effects.

Furthermore, I made use of the powerful Reducer class to manage the application state, allowing for efficient handling of complex calculations and maintaining a seamless user experience. This project has given me the opportunity to hone my programming skills and has provided me with a valuable hands-on experience in developing applications using cutting-edge technologies.

Overall, this project has further enhanced my proficiency in programming and has provided me with the necessary skills to tackle more complex and challenging projects in the future.